--- src/main/java/frameworks_and_drivers/comManager/ComManagerUser.java
@@ -1,5 +1,0 @@
-package frameworks_and_drivers.comManager;
-
-public interface ComManagerUser
-{
-}
