+++ src/main/java/frameworks_and_drivers/communication_manager/ComManagerUser.java
@@ -1,0 +1,5 @@
+package frameworks_and_drivers.communication_manager;
+
+public interface ComManagerUser
+{
+}
